Update gRPC port for testnet
From my testing, it appears that the endpoint is actually listening on port 9090. Basic testing seemed to work. One other thing that may be useful is to clarify that the endpoint is listening on plaintext gRPC instead of SSL. I'd be happy to update this PR to include `http://` if that would be acceptable.